# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Florida Lotto Scratch-off API",
    "description": "An API for the Florida Lottery Scratch-off game",
    "license": {
      "name": "Created by Kenny",
      "url": "https://www.kennywhyte.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/",
  "tags": [
    {
      "name": "Florida Lottery Scratch-off tickets"
    }
  ],

  "schemes": ["https", "http"],
  "securityDefinitions": {
    "x-auth-token": {
      "type": "apiKey",
      "name": "x-auth-token",
      "in": "header",
      "description": "A valid JWT token is needed to accessing the the POST, DELETE and UPDATE API. \n To get a JWT token, you will need to be an authenticated user."
    }
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/tickets": {
      "post": {
        "tags": ["Tickets"],
        "security": [
          {
            "x-auth-token": []
          }
        ],

        "summary": "Add a new scratch-off ticket to Database. This can only be done by an Authorized user.",
        "description": "Add a new scratch-off ticket",
        "parameters": [
          {
            "name": "ticket",
            "in": "body",
            "required": true,
            "description": "Ticket object that we want to create",
            "schema": {
              "$ref": "#/definitions/Tickets"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New ticket added",
            "schema": {
              "$ref": "#/definitions/Tickets"
            }
          },
          "401": {
            "description": "POST: Access denied. You are not authorized to perform this task.",
            "schema": {
              "$ref": "#/components/responses/UnauthorizedError"
            }
          }
        }
      },
      "get": {
        "tags": ["Tickets"],
        "summary": "Get all scratch-off tickets in the system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Tickets"
            }
          }
        }
      }
    },

    "/tickets/toptickets": {
      "get": {
        "tags": ["Tickets"],
        "summary": "Returns the \"Top 5 tickets of the Day\". The best tickets the user can buy to win the top prizes.",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Tickets"
            }
          }
        }
      }
    },

    "/tickets/{ticketId}": {
      "parameters": [
        {
          "name": "ticketId",
          "in": "path",
          "required": true,
          "description": "ID of scratch-off game that you want to locate",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Tickets"],
        "summary": "Find ticket by ID",
        "responses": {
          "200": {
            "description": "Ticket found",
            "schema": {
              "$ref": "#/definitions/Tickets"
            }
          },
          "400": {
            "description": "Invalid Id supplied"
          },
          "404": {
            "description": "Ticket not found"
          }
        }
      },
      "delete": {
        "summary": "Delete ticket with ID. This can only be done by an Authorized user.",
        "tags": ["Tickets"],
        "security": [
          {
            "x-auth-token": []
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket deleted",
            "schema": {
              "$ref": "#/definitions/Tickets"
            }
          },
          "400": {
            "description": "Invalid Id supplied"
          },
          "404": {
            "description": "Ticket not found"
          },
          "401": {
            "description": "DELETE: Access denied. You are not authorized to perform this task.",
            "schema": {
              "$ref": "#/components/responses/UnauthorizedError"
            }
          }
        }
      },
      "put": {
        "summary": "Update ticket with ID. This can only be done by an Authorized user.",
        "tags": ["Tickets"],
        "security": [
          {
            "x-auth-token": []
          }
        ],
        "parameters": [
          {
            "name": "ticket",
            "in": "body",
            "description": "Ticket with new values of properties",
            "schema": {
              "$ref": "#/definitions/Tickets"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket updated",
            "schema": {
              "$ref": "#/definitions/Tickets"
            }
          },
          "401": {
            "description": "PUT: Access denied. You are not authorized to perform this task.",
            "schema": {
              "$ref": "#/components/responses/UnauthorizedError"
            }
          }
        }
      }
    },

    "/users": {
      "post": {
        "tags": ["User"],
        "summary": "Add a new user",
        "description": "Add a new user.",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New user created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/auth": {
      "post": {
        "tags": ["User"],
        "summary": "User Authentication",
        "description": "User has to be Authenticated to POST, UPDATE and DELETE tickets",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": " User Authentication",
            "schema": {
              "$ref": "#/definitions/User Authentication"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Error logging in",
            "schema": {
              "$ref": "#/definitions/User Authentication"
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": ["User"],
        "security": [
          {
            "x-auth-token": []
          }
        ],
        "summary": "Get current user",
        "description": "Get the user who is currenlty logged in",

        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User found",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "UnauthorizedError": {
        "description": "API key is missing or invalid",
        "headers": "WWW_Authenticate",
        "schema": {
          "type": "string"
        }
      }
    }
  },
  "definitions": {
    "Tickets": {
      "properties": {
        "ticketNumber": {
          "type": "string",
          "uniqueItems": true
        },
        "ticketName": {
          "type": "string"
        },
        "topPrize": {
          "type": "string"
        },
        "topPrizeRemaining": {
          "type": "string"
        },
        "ticketCost": {
          "type": "string"
        },
        "launchDate": {
          "type": "string"
        },
        "endDate": {
          "type": "string"
        },
        "redemptionDeadline": {
          "type": "string"
        },
        "overallOddsOfWinning": {
          "type": "string"
        },
        "gameImageLink": {
          "type": "string"
        },
        "State": {
          "type": "string"
        }
      }
    },
    "User": {
      "properties": {
        "name": {
          "type": "string",
          "uniqueItems": true
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "User Authentication": {
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  }
}